Remove deprecated positioner hooks
Now they are only available in the `@remirror/react-hooks` package.
