+++ .changeset/gentle-cooks-knock.md
@@ -1,0 +1,6 @@
+---
+'@remirror/react': major
+'@remirror/react-social': patch
+---
+
+**BREAKING**: 💥 Complete move of `useMultiPositioner` and `usePositioner` to `@remirror/react-hooks`. The imports are no longer available via `@remirror/react` after being deprecated for a few weeks.
